refactor(ShipmentFilters): move into designated folder
