+++ src/pages-components/Dashboard/ShipmentsPage/ShipmentFilters/index.ts
@@ -1,0 +1,1 @@
+export { ShipmentFilters } from './ShipmentFilters';
